pyproject.toml
[project]
name = "foodpickingbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipython>=9.4.0",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.30",
    "langchain-tavily>=0.2.11",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.6",
    "langgraph-supervisor>=0.0.29",
    "python-dotenv>=1.1.1",
]
